Enable test execution on arm64
Kubebuilder tools version 1.21.2 previously used does not work on arm64
(specifically Raspberry PI). Interestingly, the latest (as of the time
of this commit) version 1.24.2 does not work either. An error message
like this causes the test to fail:

apiserver.go:43: failed to start control plane: unable to start control plane itself: failed to start the controlplane. retried 5 times: timeout waiting for process kube-apiserver to start

However, with 1.24.1 the test "mostly" passes, although it may
occasionally fail. It might be that Raspberry PI is too slow and
something times out.

Signed-off-by: Marko Vukovic <anonsoftware@gmail.com>